refactor: 컨트롤러 분류 #25
