+++ src/main/java/com/example/everguide/web/controller/TokenController.java
@@ -1,0 +1,62 @@
+package com.example.everguide.web.controller;
+
+import com.example.everguide.api.ApiResponse;
+import com.example.everguide.api.code.status.ErrorStatus;
+import com.example.everguide.api.code.status.SuccessStatus;
+import com.example.everguide.api.exception.MemberBadRequestException;
+import com.example.everguide.service.member.MemberCommandService;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequiredArgsConstructor
+public class TokenController {
+
+    private final MemberCommandService memberCommandService;
+
+    // 소셜 로그인 Refresh 토큰 쿠키 발급 후 Access 토큰 헤더 발급 위함
+    @PostMapping("/cookie-to-header")
+    public ResponseEntity<ApiResponse<String>> cookieToHeader(HttpServletRequest request, HttpServletResponse response) {
+
+        try {
+            Boolean isMember = memberCommandService.cookieToHeader(request, response);
+
+            if (isMember) {
+                return ResponseEntity.status(HttpStatus.OK)
+                        .body(ApiResponse.onSuccess(SuccessStatus._OK));
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body(ApiResponse.onFailure(ErrorStatus._UNFULL_ADDITIONAL_INFO));
+            }
+
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(ApiResponse.onFailure(ErrorStatus._NOT_FOUND, "엔티티를 찾을 수 없습니다."));
+
+        } catch (MemberBadRequestException e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(ApiResponse.onFailure(ErrorStatus._BAD_REQUEST, e.getMessage()));
+        }
+    }
+
+    // Access 토큰 만료 시 재발급
+    @PostMapping("/reissue")
+    public ResponseEntity<ApiResponse<String>> reissue(HttpServletRequest request, HttpServletResponse response) {
+
+        try {
+            memberCommandService.reissue(request, response);
+
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(ApiResponse.onSuccess(SuccessStatus._OK));
+
+        } catch (MemberBadRequestException e) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body(ApiResponse.onFailure(ErrorStatus._BAD_REQUEST, e.getMessage()));
+        }
+    }
+}
